chore: 타입 분리
